+++ examples/simple.rs
@@ -1,10 +1,11 @@
-use mvt::{Error, GeomEncoder, GeomType, Tile, Transform};
+use mvt::{Error, GeomEncoder, GeomType, Tile};
+use pointy::TransformB;
 
 fn main() -> Result<(), Error> {
     let mut tile = Tile::new(4096);
     let layer = tile.create_layer("First Layer");
     // NOTE: normally, the Transform would come from MapGrid::tile_transform
-    let b = GeomEncoder::new(GeomType::Linestring, Transform::new())
+    let b = GeomEncoder::new(GeomType::Linestring, TransformB::default())
         .point(0.0, 0.0)
         .point(1024.0, 0.0)
         .point(1024.0, 2048.0)