+++ src/encoder.rs
@@ -5,7 +5,7 @@
 //! Encoder for Mapbox Vector Tile (MVT) geometry.
 //!
 use crate::error::Error;
-use crate::geom::{Transform, Vec2};
+use pointy::{PtB, TransformB};
 
 #[derive(Copy, Clone, Debug)]
 enum Command {
@@ -42,17 +42,18 @@ pub enum GeomType {
 ///
 /// # Example
 /// ```
-/// # use mvt::{Error, GeomEncoder, GeomType, Transform};
+/// # use mvt::{Error, GeomEncoder, GeomType};
+/// # use pointy::TransformB;
 /// # fn main() -> Result<(), Error> {
-/// let geom_data = GeomEncoder::new(GeomType::Point, Transform::new())
-///                             .point(0.0, 0.0)
-///                             .point(10.0, 0.0)
-///                             .encode()?;
+/// let geom_data = GeomEncoder::new(GeomType::Point, TransformB::default())
+///     .point(0.0, 0.0)
+///     .point(10.0, 0.0)
+///     .encode()?;
 /// # Ok(()) }
 /// ```
 pub struct GeomEncoder {
     geom_tp: GeomType,
-    transform: Transform,
+    transform: TransformB,
     x: i32,
     y: i32,
     cmd_offset: usize,
@@ -66,12 +67,13 @@ pub struct GeomEncoder {
 ///
 /// # Example
 /// ```
-/// # use mvt::{Error, GeomEncoder, GeomType, Transform};
+/// # use mvt::{Error, GeomEncoder, GeomType};
+/// # use pointy::TransformB;
 /// # fn main() -> Result<(), Error> {
-/// let geom_data = GeomEncoder::new(GeomType::Point, Transform::new())
-///                             .point(0.0, 0.0)
-///                             .point(10.0, 0.0)
-///                             .encode()?;
+/// let geom_data = GeomEncoder::new(GeomType::Point, TransformB::default())
+///     .point(0.0, 0.0)
+///     .point(10.0, 0.0)
+///     .encode()?;
 /// # Ok(()) }
 /// ```
 pub struct GeomData {
@@ -104,7 +106,7 @@ impl GeomEncoder {
     ///
     /// * `geom_tp` Geometry type.
     /// * `transform` Transform to apply to geometry.
-    pub fn new(geom_tp: GeomType, transform: Transform) -> Self {
+    pub fn new(geom_tp: GeomType, transform: TransformB) -> Self {
         GeomEncoder {
             geom_tp,
             transform,
@@ -131,9 +133,9 @@ impl GeomEncoder {
 
     /// Push one point with relative coörindates.
     fn push_point(&mut self, x: f64, y: f64) {
-        let p = self.transform * Vec2::new(x, y);
-        let x = p.x as i32;
-        let y = p.y as i32;
+        let p = self.transform * PtB(x, y);
+        let x = p.x() as i32;
+        let y = p.y() as i32;
         self.data
             .push(ParamInt::new(x.saturating_sub(self.x)).encode());
         self.data
@@ -244,7 +246,7 @@ mod test {
     // Examples from MVT spec:
     #[test]
     fn test_point() {
-        let v = GeomEncoder::new(GeomType::Point, Transform::new())
+        let v = GeomEncoder::new(GeomType::Point, TransformB::default())
             .point(25.0, 17.0)
             .encode()
             .unwrap()
@@ -254,7 +256,7 @@ mod test {
 
     #[test]
     fn test_multipoint() {
-        let v = GeomEncoder::new(GeomType::Point, Transform::new())
+        let v = GeomEncoder::new(GeomType::Point, TransformB::default())
             .point(5.0, 7.0)
             .point(3.0, 2.0)
             .encode()
@@ -265,7 +267,7 @@ mod test {
 
     #[test]
     fn test_linestring() {
-        let v = GeomEncoder::new(GeomType::Linestring, Transform::new())
+        let v = GeomEncoder::new(GeomType::Linestring, TransformB::default())
             .point(2.0, 2.0)
             .point(2.0, 10.0)
             .point(10.0, 10.0)
@@ -277,7 +279,7 @@ mod test {
 
     #[test]
     fn test_multilinestring() {
-        let v = GeomEncoder::new(GeomType::Linestring, Transform::new())
+        let v = GeomEncoder::new(GeomType::Linestring, TransformB::default())
             .point(2.0, 2.0)
             .point(2.0, 10.0)
             .point(10.0, 10.0)
@@ -293,7 +295,7 @@ mod test {
 
     #[test]
     fn test_polygon() {
-        let v = GeomEncoder::new(GeomType::Polygon, Transform::new())
+        let v = GeomEncoder::new(GeomType::Polygon, TransformB::default())
             .point(3.0, 6.0)
             .point(8.0, 12.0)
             .point(20.0, 34.0)
@@ -305,7 +307,7 @@ mod test {
 
     #[test]
     fn test_multipolygon() {
-        let v = GeomEncoder::new(GeomType::Polygon, Transform::new())
+        let v = GeomEncoder::new(GeomType::Polygon, TransformB::default())
             // positive area => exterior ring
             .point(0.0, 0.0)
             .point(10.0, 0.0)