+++ src/lib.rs
@@ -11,19 +11,20 @@
 //! ## Example
 //!
 //! ```rust
-//! use mvt::{Error, GeomEncoder, GeomType, Tile, Transform};
+//! use mvt::{Error, GeomEncoder, GeomType, Tile};
+//! use pointy::TransformB;
 //!
 //! fn main() -> Result<(), Error> {
 //!     let mut tile = Tile::new(4096);
 //!     let layer = tile.create_layer("First Layer");
 //!     // NOTE: normally, the Transform would come from MapGrid::tile_transform
-//!     let b = GeomEncoder::new(GeomType::Linestring, Transform::new())
-//!                         .point(0.0, 0.0)
-//!                         .point(1024.0, 0.0)
-//!                         .point(1024.0, 2048.0)
-//!                         .point(2048.0, 2048.0)
-//!                         .point(2048.0, 4096.0)
-//!                         .encode()?;
+//!     let b = GeomEncoder::new(GeomType::Linestring, TransformB::default())
+//!         .point(0.0, 0.0)
+//!         .point(1024.0, 0.0)
+//!         .point(1024.0, 2048.0)
+//!         .point(2048.0, 2048.0)
+//!         .point(2048.0, 4096.0)
+//!         .encode()?;
 //!     let mut feature = layer.into_feature(b);
 //!     feature.set_id(1);
 //!     feature.add_tag_string("key", "value");
@@ -46,13 +47,11 @@ extern crate log;
 
 mod encoder;
 mod error;
-mod geom;
 mod mapgrid;
 mod tile;
 mod vector_tile;
 
 pub use crate::encoder::{GeomData, GeomEncoder, GeomType};
 pub use crate::error::Error;
-pub use crate::geom::{Transform, Vec2};
 pub use crate::mapgrid::{BBox, MapGrid, TileId};
 pub use crate::tile::{Feature, Layer, Tile};